frontend: move getconfig to utils config
This fixes an error related to calls in i18n/i18n.ts similar to the
circular dependency issue.

    Uncaught ReferenceError: Cannot access 'getConfig' before initialization
