+++ frontends/web/src/contexts/DarkmodeProvider.tsx
@@ -15,7 +15,7 @@
  */
 
 import { useState, useEffect, ReactNode, useCallback } from 'react';
-import { getConfig, setConfig } from '../api/backend';
+import { getConfig, setConfig } from '../utils/config';
 import { setDarkTheme, detectDarkTheme } from '../api/darktheme';
 import { runningInAndroid } from '../utils/env';
 import { useMediaQuery } from '../hooks/mediaquery';
