+++ frontends/web/src/routes/buy/exchange.tsx
@@ -34,8 +34,7 @@ import { Dialog } from '../../components/dialog/dialog';
 import { InfoButton } from '../../components/infobutton/infobutton';
 import { InfoContent } from './components/infocontent';
 import { getNativeLocale } from '../../api/nativelocale';
-import { setConfig } from '../../utils/config';
-import { getConfig } from '../../api/backend';
+import { getConfig, setConfig } from '../../utils/config';
 import { SingleValue } from 'react-select';
 import style from './exchange.module.css';
 import { CountrySelect, TOption } from './components/countryselect';
