+++ frontends/web/src/routes/buy/pocket.tsx
@@ -17,7 +17,7 @@
 import { useTranslation } from 'react-i18next';
 import { useState, useEffect, createRef } from 'react';
 import { RequestAddressV0Message, MessageVersion, parseMessage, serializeMessage, V0MessageType } from 'request-address';
-import { getConfig } from '../../api/backend';
+import { getConfig } from '../../utils/config';
 import { getTransactionList } from '../../api/account';
 import { Dialog } from '../../components/dialog/dialog';
 import { confirmation } from '../../components/confirm/Confirm';
