+++ frontends/web/src/routes/settings/electrum-servers.tsx
@@ -19,7 +19,8 @@ import { useTranslation } from 'react-i18next';
 import { ElectrumAddServer } from './electrum-add-server';
 import { ElectrumServer } from './electrum-server';
 import { TElectrumServer } from './types';
-import { getConfig, getDefaultConfig, setConfig } from '../../api/backend';
+import { getDefaultConfig } from '../../api/backend';
+import { getConfig, setConfig } from '../../utils/config';
 import { confirmation } from '../../components/confirm/Confirm';
 import style from './electrum.module.css';
 import A from '../../components/anchor/anchor';
